pyproject.toml
[tool.poetry]
name = "tutochan"
version = "0.1.0"
description = "Following channels tuto."
authors = ["Guilhem Saurel <guilhem.saurel@laas.fr>"]
license = "BSD-2-Clause"

[tool.poetry.dependencies]
python = "^3.8"
channels = "^2.4.0"
channels_redis = "^3.1.0"
psycopg2 = "^2.8.5"
python-memcached = "^1.59"
requests = "^2.24.0"
gunicorn = "^20.0.4"
ndh = "^3.10.4"
djangorestframework = "^3.11.1"
django-filter = "^2.3.0"

[tool.poetry.dev-dependencies]
ipython = "^7.18.1"
flake8 = "^3.8.3"
yapf = "^0.30.0"
selenium = "^3.141.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
